backend/requirements.txt
# Django and DRF
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
django-filter==23.2
drf-yasg==1.21.5

# Database
mysqlclient==2.1.1
PyMySQL==1.1.0

# Authentication
djangorestframework-simplejwt==5.2.2
django-allauth==0.54.0

# Utils
python-dotenv==1.0.0
Pillow==10.0.0
django-storages==1.13.2
django-redis==5.3.0
celery==5.3.1
django-celery-results==2.5.1
django-celery-beat==2.5.0

# Development
django-debug-toolbar==4.1.0
pytest-django==4.5.2
coverage==7.2.7
black==23.3.0
flake8==6.0.0
isort==5.12.0 